API設計書に基づくリファクタリング(もくもく会検索部分)
